Refactoring the code for map
