+++ maps/map1.cub
@@ -7,5 +7,5 @@ F 220, 100, 0
 C 225, 30, 0
 
 11111111
-100N1  1
+100N1001
 11111111
