+++ maps/map6.cub
@@ -6,12 +6,11 @@ SO texture/south_texture.xpm
 F 220, 100, 00
 C 225, 30, 0
 
-1111111
 1000111
 1100001
 1111111
 11
-      111111111
+      
      11000000011
     1100000000011
     10001111111111
