+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: abita <abita@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/05 15:54:05 by abita             #+#    #+#             */
-/*   Updated: 2026/02/19 22:31:34 by abita            ###   ########.fr       */
+/*   Updated: 2026/04/14 20:18:35 by abita            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,6 @@
 
 int	input_validity(t_line *line, t_data *data, int argc, char **argv)
 {
-	char			*path;
 	char			*dot;
 	char			*slash;
 	t_color_data	c_data;
@@ -30,23 +29,21 @@ int	input_validity(t_line *line, t_data *data, int argc, char **argv)
 		print_error("WARNING: too many arguments.\n");
 		ft_exit_error(data);
 	}
-	path = argv[1];
-	slash = ft_strrchr(path, '/');
-	// printf("backslash: %s\n", slash);
+	slash = ft_strrchr(argv[1], '/');
 	if (slash && slash[1] == '.')
 	{
 		printf("error: this is a hidden path\n");
 		return (EXIT_FAILURE);
 	}
-	dot = ft_strrchr(path, '.');
+	dot = ft_strrchr(argv[1], '.');
 	if (!dot || ft_strcmp(dot, ".cub") != 0)
 	{
 		print_error("ERROR: map must have '.cub' extention.\n");
 		return (EXIT_FAILURE);
 	}
-	if (parser(path, line, &c_data, &t_data) != EXIT_SUCCESS)
+	if (parser(argv[1], line, &c_data, &t_data) != EXIT_SUCCESS)
 		return (ERROR_OPENING_FILE);
-	return (0);
+	return (EXIT_SUCCESS);
 }
 
 int	main(int argc, char **argv)
