+++ src/utils/parse_utils.c
@@ -6,7 +6,7 @@
 /*   By: abita <abita@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/17 14:06:52 by abita             #+#    #+#             */
-/*   Updated: 2026/04/06 19:03:50 by abita            ###   ########.fr       */
+/*   Updated: 2026/04/14 20:23:31 by abita            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,38 +32,34 @@ int	skip_whitespace(char *line)
 		i++;
 	return (i);
 }
-int is_texture_line(char **line)
+int is_texture_line(char *line)
 {
 	int i;
-	int j;
 
 	i = 0;
-	j = 0;
-	while (line[i][j] == ' ' || line[i][j] == '\t')
+	while (line[i ] == ' ' || line[i ] == '\t')
 		i++;
-	if ((line[i][j] == 'N' && line[i][j + 1] == 'O') 
-		|| (line[i][j]== 'S' && line[i][j + 1] == 'O')
-		|| (line[i][j] == 'W' && line[i][j + 1] == 'E')
-		|| (line[i][j] == 'E' && line[i][j + 1] == 'A'))
+	if ((line[i] == 'N' && line[i  + 1] == 'O') 
+		|| (line[i]== 'S' && line[i  + 1] == 'O')
+		|| (line[i] == 'W' && line[i  + 1] == 'E')
+		|| (line[i] == 'E' && line[i + 1] == 'A'))
 	{
-		if (line[i][j + 2] == ' ' || line[i][j + 2] == '\t')
+		if (line[i + 2] == ' ' || line[i + 2] == '\t')
 			return (1);
 	}
 	return (0);
 } 
 
-int is_color_line(char **line)
+int is_color_line(char *line)
 {
 	int i;
-	int j;
 
 	i = 0;
-	j = 0;
-	while (line[i][j] == ' ' || line[i][j] == '\t')
+	while (line[i] == ' ' || line[i] == '\t')
 		i++;
-	if (line[i][j] == 'F' || line[i][j] == 'C')
+	if (line[i] == 'F' || line[i] == 'C')
 	{
-		if (line[i][j + 1] == ' ' || line[i][j + 1] == '\t')
+		if (line[i + 1] == ' ' || line[i + 1] == '\t')
 			return (1);
 	}
 	return (0);
