+++ testing/map.cub
@@ -1,3 +1,2 @@
-11111111
-1001N111
-11111111 
+100001
+111111
